Added mfg part dropdowns for ES600
Starting to implement more intelligence with awareness of the parts being reworked. Documented most common ES600 rework subjects with JSON objects.

Added a JSON file that will be the databases of MFG Part Number options

partObjects file is the module used for interfacing with the JSON database in ref_json directory